requirements.txt
python-fasthtml
click
sqlite_minutils